Modificaciones.
1) Modificación de atributos de clase LibroDTORequest, LibroDTOResponse, LibroDTOResponseUpdate.

2) Modificación de atributos de clase Libro.

3) Creación de constructor con parámetros para clase Libro.

+++ src/main/java/com/complejo/educacional/luis/durand/durand/dto/libro/LibroDTOResponse.java
@@ -38,6 +38,6 @@ public class LibroDTOResponse {
     private Long id_categoria;
     private Long id_editorial;
     private Long id_genero;
-    private Long edicion_libro;
-    private int paginas_libro;
+    private Integer edicion_libro;
+    private Integer paginas_libro;
 }
